# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Montor-ing API",
        "description": "Dokumentasi API untuk aplikasi Montor-ing (Inventory & Sales Monitoring).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Server API Utama"
        }
    ],
    "paths": {
...
        "/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Ringkasan penjualan",
                "description": "Mengambil total penjualan untuk 4 periode.",
                "operationId": "fa2272983cbe8d142dcd03ed9eee1dd1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/salesman": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Penjualan harian per salesman",
                "description": "Mengambil total penjualan masing-masing salesman untuk hari ini.",
                "operationId": "1a57f1cacecb6795c50899a489a506b7",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/salesman/yearly": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Penjualan tahunan per salesman",
                "description": "Mengambil total penjualan masing-masing salesman sepanjang tahun berjalan.",
                "operationId": "a52f8d40c6bc3ccd6bf42b5466c3dcc6",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/summary": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Ringkasan penjualan (Omzet, Laba, Margin)",
                "description": "Menghitung omzet, laba kotor, dan margin untuk 4 periode: hari ini, minggu ini, bulan ini, dan tahun ini.",
                "operationId": "892e76df1345997e0682213c6fbbb52d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/salesman-view": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard Salesman",
                "description": "Mengambil data total penjualan, return penjualan, penjualan tunai, penerimaan piutang, biaya operasional, kas di tangan dan penjualan & order 7 hari terakhir.",
                "operationId": "22d6ce3aaed8393b5b82b4dcc075b148",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/chart-monthly": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Grafik Penjualan Bulanan (Bulanan dalam 1 tahun) untuk Dashboard Web",
                "operationId": "de1865eda050dfbb14e2bdcfd7086e21",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "X-Database-Name": {
                "name": "X-Database-Name",
                "in": "header",
                "description": "Nama database yang digunakan",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari endpoint /login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}